INT20H/requirements.txt
django
gunicorn
psycopg2
django-cors-headers
apscheduler
google-api-python-client
pytz~=2023.3.post1
Pillow
django-phonenumber-field
phonenumbers
djangorestframework
django-rest-auth